requirements.txt
requests
tabulate
python-dotenv

